// packages/pqb/src/query/query.d.ts
export type QueryReturnType = QueryReturnTypeAll | 'one' | 'oneOrThrow' | 'rows' | 'pluck' | 'value' | 'valueOrThrow' | 'void';
export type QueryReturnTypeAll = undefined | 'all';
export type QueryReturnTypeOptional = 'one' | 'value';
export interface QueryHasSelect {
    __hasSelect: true;
}
export interface IsQuery {
    __isQuery: true;
}
export interface IsSubQuery {
    __subQuery: true;
}
export interface IsQueries {
    [K: string]: IsQuery;
}
export interface QueryOrExpression<T> {
    result: {
        value: Column.Pick.QueryColumnOfType<T>;
    };
}
export interface QuerySelectable {
    [K: PropertyKey]: {
        as: string;
        column: Column.Pick.QueryColumn;
    };
}
export interface Query extends IsQuery, PickQueryTable, PickQueryShape, PickQuerySelectable, QueryMethods<unknown> {
    __as: string;
    __hasSelect: boolean;
    __hasWhere: boolean;
    __defaults: EmptyObject;
    __scopes: EmptyObject;
    __defaultSelect: PropertyKey;
    result: Column.QueryColumns;
    withData: WithDataItems;
    baseQuery: Query;
    internal: QueryInternal;
    returnType: QueryReturnType;
    qb: QueryBuilder;
    columnTypes: unknown;
    __inputType: RecordUnknown;
    q: QueryData;
    then: QueryThen<unknown>;
    catch: QueryCatch;
    windows: EmptyObject;
    relations: RelationsBase;
    relationsDataForCreate: RelationsDataForCreateBase;
    relationsDataForCreateOptional: RelationsDataForCreateOptionalBase;
    relationQueries: IsQueries;
    error: new (message: string, length: number, name: QueryErrorName) => QueryError;
    __readOnly: true | undefined;
    __materialized: true | undefined;
}
export declare namespace Query {
    interface NotReadOnlyQuery extends Query {
        __readOnly: undefined;
    }
    interface MaterializedQuery extends Query {
        __materialized: true;
    }
    namespace Order {
        type Arg<T extends OrderNamespace.ArgThis> = OrderNamespace.Arg<T>;
        type Args<T extends OrderNamespace.ArgThis> = OrderNamespace.Args<T>;
    }
    namespace Pick {
        interface ReturnType {
            returnType: QueryReturnType;
        }
        interface SingleValueResult {
            result: {
                value: Column.Pick.OutputType;
            };
            returnType: 'value' | 'valueOrThrow';
        }
        interface IsNotReadOnly {
            __readOnly: undefined;
        }
        interface IsMaterialized {
            __materialized: true;
        }
    }
}
// ...
export type SetQueryReturnsRowCount<T extends PickQueryResult> = {
    [K in keyof T]: K extends 'returnType' ? 'valueOrThrow' : K extends 'result' ? {
        value: Column.Pick.QueryColumnOfType<number>;
    } : K extends 'then' ? QueryThen<number> : T[K];
};
export type SetQueryReturnsRowCountMany<T extends PickQueryResult> = {
    [K in keyof T]: K extends 'returnType' ? 'pluck' : K extends 'result' ? {
        pluck: Column.Pick.QueryColumnOfType<number>;
    } : K extends 'then' ? QueryThen<number> : T[K];
};
// ...
export declare const isQuery: (q: unknown) => q is IsQuery;