// packages/pqb/src/query/basic-features/where/where.d.ts
export interface QueryHasWhere {
    __hasWhere: true;
}
// ...
export declare const _queryWhereIn: <T>(q: T, and: boolean, arg: unknown, values: unknown[] | Iterable<unknown> | IsQuery | Expression | undefined, not?: boolean) => T & QueryHasWhere;