// packages/pqb/src/query/pick-query-types.d.ts
export interface PickQueryHasSelect {
    __hasSelect: boolean;
}
export interface PickQueryHasWhere {
    __hasWhere: boolean;
}
// ...
export interface PickQuerySelectable {
    __selectable: QuerySelectable;
}
// ...
export interface PickQueryResult {
    result: Column.QueryColumns;
}
export interface PickQueryShape {
    shape: Column.QueryColumns;
}
export interface PickQueryReturnType {
    returnType: QueryReturnType;
}
// ...
export interface PickQueryInputType {
    __inputType: RecordUnknown;
}
export interface PickQueryWithData {
    withData: WithDataItems;
}
export interface PickQueryWindows {
    windows: EmptyObject;
}
export interface PickQueryRelations {
    relations: RelationsBase;
}
// ...
export interface PickQueryAs {
    __as: string;
}